+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Allowing `-n/--no-headers` and `-d/--delimiter` flags to appear beforecolorize_main_help subcommands.
 * Adding backtick quoted strings to moonblade.
 * Adding moonblade `printf` function.
+* Adding moonblade `pad`, `lpad` & `rpad` functions.
 
 *Fixes*